#13798: Add nightly test
