+++ tests/nightly/single_card/mobilenetv2/test_mobilenetev2_modules.py
@@ -1,0 +1,1 @@
+../../../ttnn/integration_tests/mobilenetv2/test_mobilenetv2.py
